python/src/auth/requirements.txt
alembic==1.7.7
anyio==3.6.2
asgiref==3.4.1
astroid==2.11.7
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
dill==0.3.4
dnspython==2.2.1
email-validator==1.3.1
fastapi==0.83.0
greenlet==2.0.2
h11==0.13.0
httptools==0.3.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.7.0
orjson==3.6.1
parso==0.8.3
passlib==1.7.4
platformdirs==2.4.0
protobuf==3.19.6
psycopg2-binary==2.9.5
pydantic==1.9.2
PyJWT==2.4.0
pylint==2.13.9
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.46
starlette==0.19.1
tomli==1.2.3
typed-ast==1.5.4
typing_extensions==4.1.1
urllib3==1.26.14
uvicorn==0.16.0
watchgod==0.7
websockets==9.1
wrapt==1.15.0
zipp==3.6.0
